requirements.txt
telethon
requests>=2.31.0
python-dotenv
google-generativeai
